requirements.txt
fitparse 
pandas 
matplotlib 
prefect 
sqlalchemy 
psycopg2-binary 
pyyaml 
supabase
boto3
psycopg2
python-dotenv
pyarrow
s3fs